directory sheet now session based
this will reduce conflicts between people running this app. the
directory sheet id is being stored as a cookie on the browser. if the
user accessing the app does not have the cookie present, a new
directory sheet is created for them and the id is stored in their
cookies.